layoutWeight支持float类型

+++ frameworks/core/components_ng/base/view_abstract_model_ng.h
@@ -395,7 +395,7 @@ public:
 
     void SetLayoutPriority(int32_t priority) override {}
 
-    void SetLayoutWeight(int32_t value) override
+    void SetLayoutWeight(float value) override
     {
         ViewAbstract::SetLayoutWeight(value);
     }
